fixes #1962 (Tox: add icons to proto menu)

+++ protocols/Tox/src/tox_profile.cpp
@@ -89,7 +89,7 @@ bool CToxProto::LoadToxProfile(Tox_Options *options)
 		}
 		mir_free(data);
 		data = decryptedData;
-		size = decryptedSize;
+		size = (long)decryptedSize;
 	}
 
 	if (data) {
